Fix bug where byte conversions would fail at higher values than int_max
There were a few bottlenecks in the data types used. Namely:
1. `get_amt()` returned an integer
2. The `atoi()` function used to convert a string to an integer, would
   only return integers and not bigger numbers such as long (atol), or long
   long (atoll).
3. The data type in Byte was only a double and had to be upgraded to a
   long double. Similarly ints had to become at least long integers,
   etc.
4. Displaying the values also had to change its format (doubles to long double),
   (int to long int).

Note as well that these changes only apply for now to automatic size
mode. Bytesize will be modified to include these changes as well.

+++ include/bytesize.h
@@ -13,6 +13,7 @@ extern "C" {
 #include <stdbool.h>
 #include <pcre.h>
 #include <math.h>
+#include <limits.h>
 #include "log.h"
 
 #define SIZE 5
@@ -40,7 +41,7 @@ static const char *BYTE[SIZE] = {
 };
 
 typedef struct Byte {
-  double amt;
+  long double amt;
   char* unit;
   long int scaling;
 } Byte;
@@ -67,9 +68,11 @@ long int get_factor(const char *unit);
 const char *match(char *input, const char *regex);
 void display_units(const double conversion, const char* units, bool show_with_units);
 const char* get_unit(char *input);
-int get_amt(char *input);
+long int get_amt(char *input);
 double convert_units(char* input, const char* units_from, const char* units_to);
-Byte auto_size(size_t bytes, size_t scale, bool is_byte);
+//Byte auto_size(size_t bytes, size_t scale, bool is_byte);
+//Byte auto_size(u_long bytes, size_t scale, bool is_byte);
+Byte auto_size(long long int bytes, size_t scale, bool is_byte);
 
 #ifdef __cplusplus
 }
